# 🔒 Enterprise Security Enhancement - Data Security & API Optimization