--- a/requirements.txt
+++ b/requirements.txt
@@ -1,97 +1,30 @@
-# Requirements for MotmaenBash Data Repository Security Tools
+# Essential Requirements for MotmaenBash Data Repository Security Tools
 # Version: 2.0.0
 # Author: محمدحسین نوروزی (Mohammad Hossein Norouzi)
 
-# Core dependencies
+# Core dependencies for basic functionality
 aiohttp>=3.8.0
 aiofiles>=0.8.0
-asyncio-mqtt>=0.11.0
 
-# Data processing
+# Data processing (minimal)
 numpy>=1.21.0
-pandas>=1.3.0
-scipy>=1.7.0
 
-# Security and cryptography
+# Security and validation (essential only)
 cryptography>=3.4.8
-pycryptodome>=3.15.0
-bcrypt>=3.2.0
-hashlib-compat>=1.0.0
-
-# Database
-sqlite3-utils>=3.28.0
-sqlalchemy>=1.4.0
-redis>=4.0.0
-
-# HTTP and networking
-requests>=2.28.0
-urllib3>=1.26.0
-certifi>=2022.0.0
-
-# Data validation
 jsonschema>=4.0.0
-cerberus>=1.3.0
-marshmallow>=3.14.0
 
-# Monitoring and logging
-prometheus-client>=0.12.0
-structlog>=21.5.0
-python-json-logger>=2.0.0
+# HTTP and networking (essential)
+requests>=2.28.0
 
-# Performance optimization
+# Performance optimization (lightweight)
 cachetools>=4.2.0
-memory-profiler>=0.60.0
 psutil>=5.8.0
 
-# Testing
-pytest>=6.2.0
-pytest-asyncio>=0.18.0
-pytest-cov>=3.0.0
-pytest-mock>=3.6.0
-
-# Development tools
-black>=22.0.0
-flake8>=4.0.0
-mypy>=0.910
-isort>=5.10.0
-
-# Documentation
-sphinx>=4.4.0
-sphinx-rtd-theme>=1.0.0
-
 # Configuration
-pydantic>=1.9.0
 python-dotenv>=0.19.0
-click>=8.0.0
 
-# Async utilities
-asyncpg>=0.25.0
-aioredis>=2.0.0
-aiodns>=3.0.0
-
-# Compression
-lz4>=4.0.0
-zstandard>=0.17.0
-brotli>=1.0.0
-
-# Serialization
-msgpack>=1.0.0
-orjson>=3.6.0
-pickle-secure>=0.9.0
-
-# Networking utilities
-dnspython>=2.2.0
-ipaddress>=1.0.0
-netaddr>=0.8.0
-
-# Rate limiting
-ratelimit>=2.2.0
-slowapi>=0.1.0
-
-# Security scanning
-bandit>=1.7.0
-safety>=1.10.0
-
-# Monitoring
-opencensus>=0.8.0
-opencensus-ext-requests>=0.7.0
+# Testing (will be installed separately in workflow)
+# pytest>=6.2.0
+# pytest-cov>=3.0.0
+# bandit>=1.7.0
+# safety>=1.10.0