… SECURITY_SHOWCASE.md
… VERSION
… data_optimizer.py
… test-validation-report.json
… validate_data.py
… validation_report.json