@arash817 2025-07-16
This pull request will not be merged. You clearly made changes without understanding how the project works. Generating code or documentation with AI, without reading the codebase or README first, is a waste of time and not helpful to the project.